Moved BaseFragment to 'fragments' package and ProgressDialog to 'views' package within 'core' package
